fix(events): refactor Anitya events
