+++ packit_service/worker/helpers/sync_release/pull_from_upstream.py
@@ -9,8 +9,11 @@ from packit.config import JobConfig, JobType, PackageConfig
 
 from packit_service.config import ServiceConfig
 from packit_service.models import ProjectEventModel
-from packit_service.worker.events import EventData, PullRequestCommentPagureEvent
-from packit_service.worker.events.new_hotness import NewHotnessUpdateEvent
+from packit_service.worker.events import (
+    EventData,
+    NewHotnessUpdateEvent,
+    PullRequestCommentPagureEvent,
+)
 from packit_service.worker.helpers.sync_release.sync_release import SyncReleaseHelper
 
 logger = logging.getLogger(__name__)
